requirements.txt
apache-airflow==2.10.4
pyspark==3.4.3
psycopg2-binary==2.9.10
requests==2.32.3
pyyaml==6.0.2
python-telegram-bot==21.9
apache-airflow-providers-apache-spark==4.11.3